% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotateBACreads.R
\name{AnnotateBACreads}
\alias{AnnotateBACreads}
\title{Annotate BAC reads}
\usage{
AnnotateBACreads(
  blastvec,
  blastGeneA = NULL,
  blastGeneB = NULL,
  pafHost = NULL,
  minHost_mapQ = 10L,
  readLength = NULL,
  vectorSequence = NULL,
  minaln = 1L,
  MinDVDsides = 10000L
)
}
\arguments{
\item{blastvec}{Alignment of the vector on the reads. A blast result table imported with readBlast or a path to such a table.}

\item{blastGeneA}{Alignment of gene A (e.g. 18S) on the reads. A blast result table imported with readBlast or a path to such a table.}

\item{blastGeneB}{Alignment of gene B (e.g. 28S) on the reads. A blast result table imported with readBlast or a path to such a table.}

\item{pafHost}{Alignment of the reads on host genome (e.g. E Coli). A table obtained with the read_paf function or a path to a paf file (e.g. obtained with minimap2)}

\item{minHost_mapQ}{Integer. Minimum mapQ value to consider the alignment to the host genome as significant}

\item{readLength}{data.frame with 2 colums: ReadName and ReadLength. Provide directly the data frame or the path to the tab-delimited file containing the data (without header).}

\item{vectorSequence}{Either a DNAStringSet with 1 element or the path to a fasta file}

\item{minaln}{Integer. Minimum alignment length on GeneA and GeneB}

\item{MinDVDsides}{Integer. Minimum length on each side of the vector sequence to be considered a DVD read}
}
\value{
A tibble
}
\description{
The function annotates BAC reads based on:
    \itemize{
      \item{"readfq"}{Read quality}
      \item{"blastvec"}{Blast alignment of the vector on the reads}
      \item{"blastGeneA"}{Optional Blast alignment of gene A on the reads}
      \item{"blastGeneB"}{Optional Blast alignment of gene B on the reads}
      \item{"pafHost"}{Optional minimap2 alignment of the reads on the host genome}
    }
}
\examples{

# Using path to files:
## Get the path of the different files:
pathbvec <- system.file("extdata", "BAC02_BlastVector.res", package = "NanoBAC")
pathbgnA <- system.file("extdata", "BAC02_Blast18S.res", package = "NanoBAC")
pathbgnB <- system.file("extdata", "BAC02_Blast25S.res", package = "NanoBAC")
pathpaf <- system.file("extdata", "BAC02_mmap2Ecoli.paf", package = "NanoBAC")
pathRL <- system.file("extdata", "BAC02_ReadLength.tsv", package = "NanoBAC")
pathvecseq <- system.file("extdata", "VectorSequence.fa", package = "NanoBAC")

## Annotate this set of reads:
\dontrun{
myannot <- AnnotateBACreads(blastvec = pathbvec,
                            blastGeneA = pathbgnA,
                            blastGeneB = pathbgnB,
                            pafHost = pathpaf,
                            readLength = pathRL,
                            vectorSequence = pathvecseq)
}
# Import the fles first and then use them in AnnoteBACreads:
## Import the files:
blastvec = readBlast(pathbvec)
blastGeneA = readBlast(pathbgnA)
blastGeneB = readBlast(pathbgnB)
pafHost = suppressWarnings(read_paf(pathpaf))
readLength = read.table(pathRL, col.names=c("ReadName", "ReadLength"),
                        sep="\t", header = FALSE, stringsAsFactors = FALSE)
vectorSequence = Biostrings::readDNAStringSet(pathvecseq)[[1]]

## Annotate this set of reads:
\dontrun{
myannot <- AnnotateBACreads(blastvec = blastvec,
                            blastGeneA = blastGeneA,
                            blastGeneB = blastGeneA,
                            pafHost = pafHost,
                            readLength = readLength,
                            vectorSequence = vectorSequence)
}

}
