% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blaST2GR.R
\name{blaST2GR}
\alias{blaST2GR}
\title{Convert an imported blast result table to a GRanges}
\usage{
blaST2GR(blastresults, readlength)
}
\arguments{
\item{blastresults}{A tibble obtained with the readBlast function (containing strand info)}

\item{readlength}{A tibble with 2 columns named ReadName and ReadLength.
This table is expected to contain all reads, not only those having an alignment to the vector}
}
\value{
GRanges
}
\description{
Convert an imported blast result table to a GRanges
}
\examples{
## Import some data:
    Path2Blast <- system.file("extdata", "blastExample.tab", package = "NanoBAC")
    myblastResult <- readBlast(Path2Blast)
    Path2ReadLength <- system.file("extdata", "ReadLengthExample.tsv", package = "NanoBAC")
    ReadLengthTable <- read.table(Path2ReadLength,
                                  sep = "\t", header = FALSE,
                                  stringsAsFactors = FALSE)
    colnames(ReadLengthTable) <- c("ReadName", "ReadLength")
## Convert the blast result table to a GRanges
    myBlastGR <- blaST2GR(myblastResult, ReadLengthTable)
}
