% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotateBACreads.R
\name{checkBlastVar}
\alias{checkBlastVar}
\title{Check a Blast result variable}
\usage{
checkBlastVar(bvar, bform = "outfmt6")
}
\arguments{
\item{bvar}{either a path to a Blast result table (gerenated with outfmt=6) or
such a table imported with the readBlast function}

\item{bform}{character string. Currently can only be "outfmt6"}
}
\value{
a Blast result table imported with the readBlast function
}
\description{
Check a Blast result variable
}
\examples{
\donttest{
## Create a basic data frame with the right columns
df <- data.frame("QueryACC" = "MyVector",
                 "SubjectACC" = "Read1",
                 "PercentID" = "97.5",
                 "AlnLength" = 2530,
                 "NumMismatch" = 20,
                 "NumGapOpen" = 30,
                 "QueryStart" = 1,
                 "QueryEnd" = 2500,
                 "SubjectStart" = 102490,
                 "SubjectEnd" = 100000,
                 "evalue" = 0.0,
                 "bitscore" = 3000,
                 "Strand")
## checkBlastVar returns the object itself
mytab <- checkBlastVar(df)
## If we mofidy a column name
colnames(df)[1] <- "newcolname"
## then checkBlastVar throws an error
\dontrun{
mytab <- checkBlastVar(df)
}
}
}
\keyword{internal}
