% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consmat2seq.R
\name{consmat2seq}
\alias{consmat2seq}
\title{Get the consensus sequence from a consensus matrix}
\usage{
consmat2seq(x, ambiguityLetter = "N", threshold = 0.5)
}
\arguments{
\item{x}{a consensus matrix (i.e. a matrix of integers) obtained generally
using the \code{\link[Biostrings]{consensusMatrix}} function.}

\item{ambiguityLetter}{Letter used when there is an ambiguity. (Default is "N")}

\item{threshold}{\% above which a base (or a gap) is selected as the consensus}
}
\value{
A character string
}
\description{
This function is a simplified version of the \code{\link[Biostrings]{consensusString}}
function that:
\enumerate{
  \item only allows bases A, T, G, C and possibly a single ambiguity letter (default to N)
  \item does not use the IUPAC ambiguity codes in the input and output
  \item prioritizes the insertion of a gap when the \% of sequence with a gap is >\code{threshold}
  \item uses only the non gap sequences to evaluate the \% of each letter at a given position
}
Note that the function will not work if one of the base A, T, G or C or
if the gap "-" is completely absent from the alignment.
Surprisingly, the \code{\link[Biostrings]{consensusString}} function does not give identical
results when used on a \code{DNAalignment} object or a frequency matrix given
by \code{consensusMatrix} with \code{prob=TRUE}. Using the default values,
this function will give the same result as
\code{consensusString(DNAalignment-object, ambiguityMap="N", threshold=0.5)}.
}
\examples{
# Create a function to compare the consmat2seq and the consensusString functions:
ccons <- function(intmat) {
    print(Biostrings::consensusString(intmat / colSums(intmat),
                                      "N", threshold = 0.5))
    print(NanoBAC:::consmat2seq(intmat, "N", threshold = 0.5))
    }
# Create a simple matrix:
consmat <- matrix(c(10L, 0L, 0L, 0L, 0L, 0L,
                    0L, 10L, 0L, 0L, 0L, 0L,
                    0L, 0L, 10L, 0L, 0L, 0L,
                    0L, 0L, 0L, 10L, 0L, 0L),
                    nrow = 6,
                    dimnames = list(c("A", "T", "G", "C", "N", "-")))
ccons(consmat) # same result
consmat[,3] <- c(4L, 0L, 0L, 0L, 0L, 6L)
ccons(consmat) # same
consmat[,3] <- c(4L, 1L, 0L, 0L, 0L, 5L)
ccons(consmat) # same
consmat[,3] <- c(5L, 0L, 0L, 0L, 0L, 5L)
ccons(consmat) # different. (favor the gap)
consmat[,3] <- c(4L, 2L, 0L, 0L, 0L, 4L)
ccons(consmat) #different (only consider letters)
consmat[,3] <- c(4L, 2L, 1L, 1L, 0L, 2L)
ccons(consmat) #different (only consider letters)
consmat[,3] <- c(4L, 2L, 1L, 1L, 1L, 1L)
ccons(consmat) #same

}
\keyword{internal}
