% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterBACreads.R
\name{FilterBACreads}
\alias{FilterBACreads}
\title{Select specific reads from a table of BAC read annotations obtained by the function \code{AnnotateBACreads}}
\usage{
FilterBACreads(
  ReadClass = NULL,
  readtype = "VDV",
  MinReadLength = NULL,
  alnGeneA = NULL,
  alnGeneB = NULL,
  isHostAlign = NULL,
  VDVInsertLengthTarget = NULL,
  VDVInsertLengthTolerance = 0.05
)
}
\arguments{
\item{ReadClass}{Either a tibble obtained with the \code{\link{AnnotateBACreads}} function or
a path to an RDS file containing such a tibble}

\item{readtype}{Character vector. Type(s) of reads to select.
Values should be in c("VDV", "DVD", "VD", "D", "V", "Chimeric", NA).
(Defaults to "VDV")}

\item{MinReadLength}{Integer. Minimum length of the selected reads. Reads below this length are dropped.
(defaults to NULL, i.e. no filtering on read length)}

\item{alnGeneA}{Logical. Should the selected read align to GeneA ? (defaults to NULL, i.e no filtering on GeneA alignment)}

\item{alnGeneB}{Logical. Should the selected read align to GeneB ? (defaults to NULL, i.e no filtering on GeneB alignment)}

\item{isHostAlign}{Logical. Should the selected reads align to the host genome (TRUE) or not (FALSE)?
(defaults to NULL, i.e no filtering on host genome alignment alignment)}

\item{VDVInsertLengthTarget}{Integer. What is the estimated length of the DNA insert (without the vector) in the VDV reads to be selected?
(Defaults to NULL, no filtering on VDV read length)}

\item{VDVInsertLengthTolerance}{Numeric. Number in [0,1[. What tolerance (in \%) is accepted around VDVInsertLengthTarget?}
}
\value{
tibble with selected reads.
}
\description{
Select specific reads from a table of BAC read annotations obtained by the function \code{AnnotateBACreads}
}
\details{
The function uses the column "LongestDNA" to filter for the size of VDV reads
}
\examples{
## Path to file (.rds) created with the AnnotateBACreads function
pathRC <- system.file("extdata", "BAC02_ReadClass.rds", package = "NanoBAC")
## Extract the annotation of all VDV reads longer than 111 kb
FilterBACreads(pathRC, "VDV", 111e3)
## Extract the annotation of the VDV reads which have an insert that is 104 kb long (+/- 0.1\%)
FilterBACreads(pathRC, "VDV",
               VDVInsertLengthTarget = 104e3,
               VDVInsertLengthTolerance = 0.001)
## Extract the annotation of the D reads that map to the E. coli genome
FilterBACreads(pathRC, "D", isHostAlign = TRUE)
## Extract the annotation of the VD and the VDV reads that align to both GeneA and GeneB
FilterBACreads(pathRC, c("VD", "VDV"), alnGeneA = TRUE, alnGeneB = TRUE)
## There are no VDV reads that align to the E. coli host genome
\dontrun{
FilterBACreads(pathRC, "VDV", isHostAlign = TRUE)
}
}
