% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findVDVjunctions.R
\name{FindVDVjunctions}
\alias{FindVDVjunctions}
\title{Find the vector-insert junctions at the beginning and end of a VDV read}
\usage{
FindVDVjunctions(
  ReadName = NULL,
  ReadDNA,
  ReadVecAlign,
  RestrictionSite = "G^AATTC",
  VectorSequence = NULL,
  UnalignedVectorLength = 1000L,
  SideSeqSearch = 10L
)
}
\arguments{
\item{ReadName}{character string. Name of the read}

\item{ReadDNA}{A DNAString or DNAStringSet with the read sequence}

\item{ReadVecAlign}{Table with the Blast results from aligning th evector on the read}

\item{RestrictionSite}{Character string in the for "G^AATTC" indicating the sequence and the cut site}

\item{VectorSequence}{A DNAString or DNAStringSet of length 1 with the vector sequence starting and ending
with the full sequence of the restriction site used for cloning}

\item{UnalignedVectorLength}{Integer. If more than \code{UnalignedVectorLength}bp of expected vector sequence
is not correctly aligned at the vector-insert junction,
the function will return a message}

\item{SideSeqSearch}{Integer. If the expected restriction site is not found at a vector-insert junction
then the algorithm will try to search for the vector sequence of length
\code{SideSeqSearch} bp that is adjacent to the restriction site
(Default to 10 bp)}
}
\value{
a named character vector with the read name, the read strand and the coordinates of the first and last bases of the insert
}
\description{
Find the vector-insert junctions at the beginning and end of a VDV read
}
\examples{
# Some dummy sequences (use paste0 for clarity / comparison of sequences):
## Vector sequence start and ends with the HindIII site used for cloning ("A^AGCTT")
vector <- Biostrings::DNAString(paste0(
            "AAGCTTTATTAAGACACCCGGTATGCTTCAGGATCGTTCGGACTAA",
            "ACCGTAACTGCGATATTTTAGGCGTGTTACAAGCTT"))
read <- Biostrings::DNAString(paste0(
            "ACCGTAACTGCGATATTTTAGGCGTGTTACAAGCTT",
            "GCTAGATCGCGCGATATGTG",
            "AAGCTTTATTAAGACACCCGGTATGCTTCAGGATCGTTCGGACTAA"))
noisyread <- Biostrings::DNAString(paste0(
            "ACCGTAACTGCGTTTTTTTAGGCGTGTTACAAGCTT",
            "GCTAAATCGCGCGCTATGTG",
            "GGGCTTTATTAAGACACCCGGTATGCTTTCAGGATCGTTCGGACTAA"))
# Import the blastn results for these sequences (alignment of vector on the reads):
readaln <- readBlast(system.file("extdata",
                                 "juncEx_vec_read.res",
                                 package = "NanoBAC"))
noisyreadaln <- readBlast(system.file("extdata",
                                      "juncEx_vec_noisyread.res",
                                      package = "NanoBAC"))
# Get the coordinates of the insert sequence:
FindVDVjunctions("read", read, readaln, "A^AGCTT", vector)
# With the noisy read, the restriction site is not found but an adjacent sequence is:
FindVDVjunctions("noisyread", noisyread, noisyreadaln, "A^AGCTT", vector)
}
