% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDVDnames.R
\name{getDVDnames}
\alias{getDVDnames}
\title{Identify DVD reads based on alignment of the vector on the reads}
\usage{
getDVDnames(
  alignGR,
  vectorLength,
  PercentVecLength = 0.95,
  MinDNASides = 10000L
)
}
\arguments{
\item{alignGR}{a \code{GRanges} object containing the alignment of the vector on the reads}

\item{vectorLength}{integer Length of the vector (in bp)}

\item{PercentVecLength}{numeric in [0.5,1] indicating the \% of the vector length that must be aligned on the read to be considered a DVD read}

\item{MinDNASides}{minimum length of non-vector DNA on each side of the central vector alignment}
}
\value{
a character vector with the names of DVD reads
}
\description{
DVD reads contain DNA-vector-DNA
  DVD reads have a single vector alignment covering at least PercentVecLength\% (default 95\%) of the vector length.
}
\examples{
## Create a GRanges. only Read1 and Read2 are DVD reads
## vector is 10kb
## alignment of vector on Read2 covers 9501 bp (>95\\% of vector length)
  rgr <- GenomicRanges::GRanges(c("Read1:12e3-22e3",
                                  "Read2:20e3-29.6e3",
                                  "Read3:1-2000", "Read3:98001-1e5"),
                                seqlengths = c("Read1"=5e4, "Read2"=6e4,
                                               "Read3"=1e5, "Read4"=5e4),
                                 QueryRange.width = c(1e4, 9501, 2000, 2000))
## Names of VDV reads (using 95\\% of vector length as threshold):
  getDVDnames(rgr, 10000, 0.95)
## With 98\\% of vector length only Read1 will be selected as a DVD read
  getDVDnames(rgr, 10000, 0.98)
## If requiring at least 15kb on each side of the vector then only Read2 is DVD
  getDVDnames(rgr, 10000, 0.95, 15e3)
}
