% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVDVnames.R
\name{getVDVnames}
\alias{getVDVnames}
\title{Identify VDV reads based on alignment of the vector on the reads}
\usage{
getVDVnames(alignGR, vectorLength, minReadLength = 2000, SideWidth = 1000)
}
\arguments{
\item{alignGR}{a \code{GRanges} object containing the alignment of the vector on the reads}

\item{vectorLength}{integer Length of the vector (in bp)}

\item{minReadLength}{integer. Minimum read length to be a VDV read (defaults to 2kb)}

\item{SideWidth}{integer. Length on each side of the read that must align to some extent to the vector}
}
\value{
a character vector of names of the VDV reads
}
\description{
VDV reads contain vector-DNA-vector sequences.
  Only reads longer than minReadLength (>=2kb) can be annotated as VDV reads.
  VDV reads contain an alignment of the vector in their first 1kb and in their last 1kb
  but do not contain any alignment to the vector in the center region (i.e. excluding 1.15x the length of the vector on each side)
}
\examples{
## Create a GRanges. only Read1 and Read2 are VDV reads
  rgr <- GenomicRanges::GRanges(c("Read1:1-2000", "Read1:98001-1e5",
                                  "Read2:100-1800", "Read2:99e3-1e5",
                                  "Read3:1e4-1.4e4"),
                                seqlengths = c("Read1"=1e5, "Read2"=1e5,
                                               "Read3"=2e4, "Read4"=5e4))
## Names of VDV reads:
  getVDVnames(rgr, 4000)
}
