% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeVarseq.R
\name{makeVarseq}
\alias{makeVarseq}
\title{Introduce variations in a DNA sequence}
\usage{
makeVarseq(
  dnaseq,
  lettrs = c("A", "T", "G", "C"),
  subst = c(0.03, 0.06),
  del = c(0.01, 0.05),
  ins = c(0.01, 0.05),
  returnString = TRUE
)
}
\arguments{
\item{dnaseq}{Either a character string, a DNAString or a character vector with individual characters.
The DNA sequence in which variations are introduced}

\item{lettrs}{character vector}

\item{subst}{numeric vector of length 2 with values in [0,1].
Percentage of substitutions (upper and lower bounds).
Default is 3-6\% substition.}

\item{del}{numeric vector of length 2 with values in [0,1].
Percentage of deletions (upper and lower bounds).
Default is 1-5\% deletion.}

\item{ins}{numeric vector of length 2 with values in [0,1].
Percentage of insertions (upper and lower bounds).
Default is 1-5\% insertions.}

\item{returnString}{Logical. Should the function return a single character string?}
}
\value{
either a vector of individual characters
        or a single character string if \code{returnString==TRUE}
}
\description{
Introduce variations in a DNA sequence
}
\examples{
set.seed(12345) # for reproducibility
## Introduce ~10\% variation in a sequence
makeVarseq("ATGCATGCATGCATGCATGCATGC",
           subst = c(0.08, 0.12),
           del = c(0.08, 0.12),
           ins = c(0.08, 0.12))
}
