% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paf2gr.R
\name{paf2gr}
\alias{paf2gr}
\title{Convert an paf file imported with \code{\link[NanoBAC]{read_paf}} to a \code{GRanges}}
\usage{
paf2gr(paf, focus = c("target", "reads"), includetags = TRUE)
}
\arguments{
\item{paf}{A tibble with a paf file imported with the read_paf function}

\item{focus}{Character string. Either "target" or "reads" indicating which ranges will be represented in the GRanges}

\item{includetags}{Logical. Should the tags in the paf file be present in mcols from the GRanges (default is TRUE)?}
}
\value{
a \code{GRanges} object
}
\description{
the ranges in the resulting \code{GRanges} object correspond either to the reads or to the target depending on the value of the focus argument
}
\examples{
## Example data set:
    Path2paf <- system.file("extdata", "BAC02_mmap2Ecoli.paf", package = "NanoBAC")
## Import the data (ignore the parsing failures):
   suppressWarnings(mypaf <- read_paf(Path2paf))
## Convert the dataset to a GRanges focusing on the reads
   mypafgr <- paf2gr(mypaf)

}
\seealso{
\code{\link[NanoBAC]{read_paf}}
}
