% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBlast.R
\name{readBlast}
\alias{readBlast}
\title{Read a blast result file
  (currently only for files created using the blast parameter outfmt=6)}
\usage{
readBlast(blastpath, blastformat = "outfmt6")
}
\arguments{
\item{blastpath}{character string. Path to a Blast result file}

\item{blastformat}{format of the Blast result file (currently only the default outfmt=6 is supported)}
}
\value{
a tibble
}
\description{
Read a blast result file
  (currently only for files created using the blast parameter outfmt=6)
}
\examples{
## Example data set:
    Path2Blast <- system.file("extdata", "BAC02_Blast18S.res", package = "NanoBAC")
## Import the data:
   myblastResult <- readBlast(Path2Blast)

}
