% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectSingularBlastALN.R
\name{SelectSingularBlastALN}
\alias{SelectSingularBlastALN}
\title{Filter a blast result table for alignment overlapping other alignments}
\usage{
SelectSingularBlastALN(aln, rl, threshold = 0.5)
}
\arguments{
\item{aln}{A tibble obtained with the \code{\link[NanoBAC]{readBlast}} function (containing strand info)}

\item{rl}{A ReadLength table, i.e. a data frame with 2 columns: ReadName and ReadLength}

\item{threshold}{Number in ]0,1]. Remove all alignments that overlap with other alignments on > threshold \% of their length}
}
\value{
a character vector with the row numbers corresponding to alignments to keep
    (i.e. that don't overlap on >threshold% of their length with other alignments)
}
\description{
Filter a blast result table imported with \code{\link[NanoBAC]{readBlast}}
    to remove alignments with large overlaps (e.g. > 50\%) with other alignments
}
\examples{
## Example dataset. The vector has a repetitive region that systematically give multiple alignments
  Path2OVL <- system.file("extdata", "BAC02_BlastVector.res", package = "NanoBAC")
  alignmt <- readBlast(Path2OVL)
  nrow(alignmt)
## Read lengths
  Path2ReadLength <- system.file("extdata", "BAC02_ReadLength.tsv", package = "NanoBAC")
  ReadLengthTable <- read.table(Path2ReadLength,
                                sep = "\t", header = FALSE,
                                stringsAsFactors = FALSE,
                                col.names = c("ReadName", "ReadLength"))
## Filter the data to keep alignments not overlapping on more than 50\% of their length
  filtaln <- alignmt[SelectSingularBlastALN(alignmt, ReadLengthTable),]
  nrow(filtaln)  #many alignments removed

}
\references{
The code for this function is based on a suggestion by Michael Lawrence: \url{https://support.bioconductor.org/p/72656/}
}
\seealso{
\link{readBlast}
}
\author{
Pascal GP Martin
}
