% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectVDVreads.R
\name{selectVDVreads}
\alias{selectVDVreads}
\title{Select VDV reads from a set of Nanopore reads}
\usage{
selectVDVreads(
  ReadClass = NULL,
  SizeTolerance = 0.05,
  WithGeneA = NULL,
  WithGeneB = NULL,
  MaxClusters = 10L,
  makePlot = TRUE,
  plotVar = c("ReadLength", "InsertLength")
)
}
\arguments{
\item{ReadClass}{Either a tibble obtained with the \code{\link{AnnotateBACreads}} function
or a path to an rds file containing such a file}

\item{SizeTolerance}{A single number in [0,1[.
Reads with a size corresponding to the estimated size +/- \code{SizeTolerance}\% are selected}

\item{WithGeneA}{Logical. Should the VDV reads align with GeneA? Default is NULL, i.e. no filtering on GeneA alignment}

\item{WithGeneB}{Logical. Should the VDV reads align with GeneB? Default is NULL, i.e. no filtering on GeneB alignment}

\item{MaxClusters}{Integer. Maximum number of clusters to make with VDV reads to estimate the size of the DNA insert.
Default to 10 which in our hands works in most situations.}

\item{makePlot}{Logical. Should a plot be produced (and saved in the result)}

\item{plotVar}{Character string. Variable to plot: either \code{'ReadLength'} or \code{InsertLength}.}
}
\value{
A list with the following objects:
\itemize{
  \item{VDVreads}{A tibble with info on VDV reads: name, length, selection by the procedure, cluster}
  \item{InsertSizeEstimate}{A list with result from the \code{\link{estimateBACsizeFromVDV}} function}
  }
if makePlot is TRUE, then the list also contains:
\itemize{
  \item{VDVlengthPlot}{The plot (ggplot2 object)}
  \item{PlotType}{The variable used to make the plot, i.e. the plotVar argument that was used to create the object}
  \item{xcoord}{The x coordinates of the points (obtained using geom_jitter) in order to reproduce the plot easily}
}
}
\description{
The function does the following:
\itemize{
  \item{Selects VDV reads}{This is done using \code{\link{FilterBACreads}}}
  \item{Estimate BAC size}{The size of the insert sequence is estimated using \code{\link{estimateBACsizeFromVDV}}}
  \item{Select VDV reads of the right size}{VDV reads within +/- \code{SizeTolerance}\% of the estimated size are selected}
  \item{Plot VDV read size}{The plot illustrates the size of all VDV reads and the selection process}
  }
By default, if alignemnt to the host genome is provided in the \code{ReadClass} object (column \code{HostAlign}),
then the selected VDV reads do not align to the host genome
}
\examples{
## Path to file (.rds) created with the AnnotateBACreads function
pathRC <- system.file("extdata", "BAC02_ReadClass.rds", package = "NanoBAC")
## Import the data
annotatedReads <- readRDS(pathRC)
## Select VDV reads that contain alignment to GeneA and GeneB
myVDVreads <- selectVDVreads(ReadClass = annotatedReads,
                             WithGeneA = TRUE,
                             WithGeneB = TRUE)
## Same but making less clusters and producing a plot on insert length rather than read length
myVDVreads <- selectVDVreads(ReadClass = annotatedReads,
                             WithGeneA = TRUE,
                             WithGeneB = TRUE,
                             MaxClusters = 8,
                             plotVar = "InsertLength")
}
\author{
Pascal GP Martin
}
