% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitDVDreads.R
\name{splitDVDreads}
\alias{splitDVDreads}
\title{Select DVD reads from a set of Nanopore reads and split the reads in 2 parts both containing the vector (DV and VD)}
\usage{
splitDVDreads(
  ReadClass = NULL,
  blastvec = NULL,
  FastaFile = NULL,
  WithGeneA = NULL,
  WithGeneB = NULL,
  MinDNAlength = 10000L
)
}
\arguments{
\item{ReadClass}{Either a tibble obtained with the \code{\link{AnnotateBACreads}} function
or a path to an rds file containing such a file}

\item{blastvec}{Either a table imported with \code{\link{readBlast}} or a path to a blast file
obtained by aligning th evectors on the reads and using \code{-outfmt 6}}

\item{FastaFile}{Either a \code{DNAStringSet} object containing the full read sequences
or a path to a fasta file containing these sequences}

\item{WithGeneA}{Logical. Should the VDV reads align with GeneA? Default is NULL, i.e. no filtering on GeneA alignment}

\item{WithGeneB}{Logical. Should the VDV reads align with GeneB? Default is NULL, i.e. no filtering on GeneB alignment}

\item{MinDNAlength}{Integer. Minimum length of the DNA fragment to keep the reads in the results}
}
\value{
A list with:
    \itemize{
        \item{ReadDefinition}{ a \code{DNAStringSet} with the split reads}
        \item{ReadSequence}{ a \code{GRanges} object with the definition of the DV/VD reads}
    }
    Note that reads with alignment on the opposite strand of the vector ("-" strand)
    are automatically reverse complemented
}
\description{
The function does the following:
\itemize{
  \item{Selects DVD reads}{This is done using \code{\link{FilterBACreads}}}
  \item{Split the read sequence in DV and VD}{Based on vector alignment, split the read sequence in DV and VD}
  \item{Filter based on size}{Keep only the split reads with a DNA fragment that is at least \code{MinDNAlength}bp long}
  \item{reverse complement reads on minus strand}{strand is determined based on vector alignment}
  \item{Return the split reads}{The split reads are returned as a DNAString object}
  }
By default, if alignemnt to the host genome is provided in the \code{ReadClass} object (column \code{HostAlign}),
then the selected DVD reads are selected to not show any significant alignment to the host genome
}
\examples{
## For simplicity (and to limit file size) we only keep the data for 5 pre-selected DVD reads
## Path to file (.rds) created with the AnnotateBACreads function
pathRC <- system.file("extdata", "BAC02_ReadClass.rds", package = "NanoBAC")
RC <- readRDS(pathRC)
selectedReads <- c("BAC02R5572", "BAC02R21438", "BAC02R1152",
                   "BAC02R20794", "BAC02R6278" )
RC <- RC[RC$ReadName \%in\% selectedReads,]
## Path to a fasta file containing the sequence of the 5 DVD reads
pathFasta <- system.file("extdata", "BAC02_5DVDreads.fa", package = "NanoBAC")
## Path to the file containing the result from the Blast alignment of the vector on the reads
pathBlast <- system.file("extdata", "BAC02_BlastVector.res", package = "NanoBAC")
## Select DVD reads and split the reads
myDVDreads <- splitDVDreads(ReadClass = RC,
                            blastvec = pathBlast,
                            FastaFile = pathFasta,
                            WithGeneA = TRUE,
                            WithGeneB = TRUE,
                            MinDNAlength = 35000)
## Read sequences:
myDVDreads$ReadSequence
## Read definitions:
myDVDreads$ReadDefinition
}
\author{
Pascal GP Martin
}
