% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_paf.R
\name{tidy_paf_tags}
\alias{tidy_paf_tags}
\title{Tidy the flag columns imported from a paf file}
\usage{
tidy_paf_tags(Somedata)
}
\arguments{
\item{Somedata}{A \code{tibble} as imported during the [NanoBAC::read_paf()] function}
}
\value{
tibble
}
\description{
Tidy the flag columns imported from a paf file
}
\details{
This function is not intended to be used directly but is used in the [NanoBAC::read_paf()] function
}
