% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensusFromMSF.R
\name{consensusFromMSF}
\alias{consensusFromMSF}
\title{Obtain the consensus sequence from an MSF file}
\usage{
consensusFromMSF(
  msfALN = NULL,
  gapthreshold = 0.5,
  removegaps = TRUE,
  consname = "cons"
)
}
\arguments{
\item{msfALN}{Character string. Path to an MSF file obtained with a multiple aligment program}

\item{gapthreshold}{Numeric in ]0,1[. (Default is 0.5)}

\item{removegaps}{Logical. Should the gaps be removed in the output sequence? (Default is TRUE)}

\item{consname}{Character string. Name of the consensus sequence. (Default is "cons")}
}
\value{
A DNAStringSet of length 1 with the consensus sequence
}
\description{
Obtain the consensus sequence from an MSF file
}
\examples{

}
